+++ Sources/RichTextKit/Context/RichTextContext+Actions.swift
@@ -12,7 +12,12 @@ public extension RichTextContext {
     
     /// Handle a certain rich text action.
     func handle(_ action: RichTextAction) {
-        triggerAction = action
+        switch action {
+        case .setAlignment(let align): textAlignment = align
+        case .stepFontSize(let points): fontSize += CGFloat(points)
+        case .toggleStyle(let style): toggle(style)
+        default: triggerAction = action
+        }
     }
 
     /// Check if the context can handle a certain action.
@@ -22,10 +27,12 @@ public extension RichTextContext {
         case .dismissKeyboard: return true
         case .print: return false
         case .redoLatestChange: return canRedoLatestChange
+        case .setAlignment: return true
         case .stepFontSize: return true
         case .stepIndent(let points):
             return points < 0 ? canDecreaseIndent : canIncreaseIndent
         case .stepSuperscript: return false
+        case .toggleStyle: return true
         case .undoLatestChange: return canUndoLatestChange
         }
     }